2. This report, by its very length, defends itself against the risk of being read

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,13 +7,12 @@
  * Return: length.
  */
 
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
-	int len = 0;
+	size_t len = 0;
 
 	while (*str++)
 		len++;
 
 	return (len);
 }
-
